restructure directories for graphql and rest api
