+++ templates/kubernetes/overlays/staging/kustomization.yml
@@ -18,4 +18,4 @@ configMapGenerator:
     - ENVIRONMENT=staging
     - DOMAIN=<% index .Params `stagingHostRoot` %>
     - S3_BUCKET=files.<% index .Params `stagingHostRoot` %>
-    - GRAPHQL_BASE_URL=https://<% index .Params `stagingBackendSubdomain` %><% index .Params `stagingHostRoot` %>
+    <%if eq (index .Params `apiType`) "graphql" %>- GRAPHQL_BASE_URL=https://<% index .Params `stagingBackendSubdomain` %><% index .Params `stagingHostRoot` %><% end %>
