+++ templates/src/app.js
@@ -4,13 +4,13 @@ const morgan = require("morgan");
 const dbDatasource = require("./db");
 <%if eq (index .Params `apiType`) "graphql" %>const { ApolloServer } = require("apollo-server-express");
 
-const typeDefs = require("./app/graphql/schema");
-const { createStore } = require("./app/graphql/initdb");
-const resolvers = require("./app/graphql/resolvers");
-const LaunchAPI = require("./app/graphql/datasources/LaunchAPI");
-const UserDB = require("./app/graphql/datasources/UserDB");
-const FileAPI = require("./app/graphql/datasources/FileAPI");
-const AuthAPI = require("./app/graphql/datasources/AuthAPI");<% end %>
+const typeDefs = require("./graphql/schema");
+const { createStore } = require("./graphql/initdb");
+const resolvers = require("./graphql/resolvers");
+const LaunchAPI = require("./graphql/datasources/LaunchAPI");
+const UserDB = require("./graphql/datasources/UserDB");
+const FileAPI = require("./graphql/datasources/FileAPI");
+const AuthAPI = require("./graphql/datasources/AuthAPI");<% end %>
 <%if eq (index .Params `fileUploads`) "yes" %>const fileRoutes = require("./app/file");<% end %>
 const statusRoutes = require("./app/status");
 <%if eq (index .Params `userAuth`) "yes" %>const authRoutes = require("./app/auth");<% end %>
