+++ templates/src/app/status/index.js
@@ -3,16 +3,17 @@ var { Router } = require("express");
 var router = Router();
 
 router.get("/ready", (req, res) => {
-  res.send("OK");
+  res.json({ ready: "OK" });
 });
 
 router.get("/alive", (req, res) => {
-  res.send("OK");
+  res.json( {alive: "OK"} );
 });
 
 router.get("/about", (req, res) => {
-  res.send({
-    podName: process.env.POD_NAME,
+  var podName = (process.env.POD_NAME)?process.env.POD_NAME:"zero-node-backend";
+  res.json({
+    podName: podName,
   });
 });
