+++ templates/src/graphql/datasources/AuthAPI.js
