+++ templates/src/graphql/initdb.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require('sequelize');
-const db = require("../../db");
+const db = require("../db");
 
 module.exports.createStore = () => {
   const users = db.define('user', {
